+++ routes/books.js
@@ -1,13 +1,17 @@
 const express = require("express")
+const isAuth = require("../middleware/is-auth");
 const router = express.Router()
 // const multer = require('multer')
 const Book = require("../models/book")
 const  Author = require("../models/author")
-const author = require("../models/author")
+const User = require("../models/user")
 const imageMimeTypes = ['image/jpeg', 'image/png', 'image/gif', 'image/jpg']
 //all authors route
-router.get('/',async (req,res)=>{
-    let query = Book.find()
+router.get('/', isAuth, async (req,res)=>{
+    const email = req.session.email;
+    const user = await User.find({email: email})
+    let searchOptions = {user : user}
+    let query = Book.find(searchOptions)
     if(req.query.title != null && req.query.title != '')
         query = query.regex('title' , new RegExp(req.query.title , 'i'))
     if(req.query.publishedBefore != null && req.query.publishedBefore != '')
@@ -25,32 +29,47 @@ router.get('/',async (req,res)=>{
         res.redirect('/')
     }
 });
-// new author route
-router.get('/new',async(req,res)=>{
-    renderNewPage(res, new Book())
+// new Book route
+router.get('/new',isAuth,async(req,res)=>{
+    renderNewPage(req, res, new Book())
 });
 router.get('/:id',async(req,res)=>{
     try{
         const book = await Book.findById(req.params.id).populate('author').exec()
+        const user = await User.findById(book.user)
+        if(req.session.email != user.email) {
+            res.redirect('/')
+            return
+        }
         res.render('books/show' , {book: book})
     }catch{
         res.redirect('/')
     }
 });
 //Edit book route
-router.get('/:id/edit', async(req,res)=>{
+router.get('/:id/edit', isAuth,async(req,res)=>{
     try{
         const book = await Book.findById(req.params.id)
-        renderEditPage(res, book)
+        const user = await User.findById(book.user)
+        if(req.session.email != user.email) {
+            res.redirect('/')
+            return
+        }
+        renderEditPage(req, res, book)
     }catch{
         res.redirect('/')
     }
 });
 //Update book
-router.put('/:id', async (req, res)=>{
+router.put('/:id', isAuth,async (req, res)=>{
     let book
     try{
         book = await Book.findById(req.params.id)
+        const user = await User.findById(book.user)
+        if(req.session.email != user.email) {
+            res.redirect('/')
+            return
+        }
         book.title = req.body.title
         book.author = req.body.author
         book.publishDate =  new Date(req.body.publishDate)
@@ -60,23 +79,25 @@ router.put('/:id', async (req, res)=>{
             saveCover(book , req.body.cover)
         }
         await book.save()
-        res.redirect(`/books/${book.id}`)
+        res.redirect(`/files/books/${book.id}`)
     }
     catch(error){
         console.log(error)
         if(book == null){
             res.redirect('/')
         }else{
-            renderEditPage(res, book, true)
+            renderEditPage(req, res, book, true)
         }
     }
 })
 // delete all books
-router.delete('/deleteAll', async (req, res)=>{
+router.delete('/deleteAll', isAuth,async (req, res)=>{
+    const email = req.session.email;
+    const user = await User.findOne({email: email})
     let book
     try{
-        await Book.deleteMany({})
-        res.redirect('/books')
+        await Book.deleteMany({user: user})
+        res.redirect('/files/books')
     }
     catch(error){
         console.log(error)
@@ -86,15 +107,20 @@ router.delete('/deleteAll', async (req, res)=>{
                 errorMessage: 'Could not remove Books'
             })
         }else{
-            res.redirect('/books')
+            res.redirect('/files/books')
         }
     }
 })
 //delete book
-router.delete('/:id', async (req, res)=>{
+router.delete('/:id', isAuth,async (req, res)=>{
     let book
     try{
         book = await Book.findById(req.params.id)
+        const user = await User.findById(book.user)
+        if(req.session.email != user.email) {
+            res.redirect('/')
+            return
+        }
         await book.deleteOne()
         res.redirect('/books')
     }
@@ -111,13 +137,15 @@ router.delete('/:id', async (req, res)=>{
     }
 })
 //new author route
-router.post('/', async (req,res)=>{
+router.post('/', isAuth,async (req,res)=>{
+    const email = req.session.email
     const book = new Book({
         title: req.body.title,
         author:req.body.author,
         publishDate: new Date(req.body.publishDate),
         pageCount: req.body.pageCount,
-        description: req.body.description
+        description: req.body.description,
+        user: await User.findOne({email: email})
     })
     saveCover(book , req.body.cover)
     try{
@@ -126,19 +154,21 @@ router.post('/', async (req,res)=>{
     }
     catch(error){
         console.log(error)
-        renderNewPage(res, book, true)
+        renderNewPage(req, res, book, true)
     }
 })
 
-async function renderNewPage(res, book, hasError = false){
-    renderFormPage(res, book, 'new' , hasError)
+async function renderNewPage(req, res, book, hasError = false){
+    renderFormPage(req, res, book, 'new' , hasError)
 }
-async function renderEditPage(res, book, hasError = false){
-    renderFormPage(res, book, 'edit' , hasError)
+async function renderEditPage(req, res, book, hasError = false){
+    renderFormPage(req, res, book, 'edit' , hasError)
 }
-async function renderFormPage(res, book, form, hasError = false){
+async function renderFormPage(req, res, book, form, hasError = false){
+    const email = req.session.email
     try{
-        const authors = await Author.find({})
+        const user = await User.findOne({ email: email })
+        const authors = await Author.find({user: user})
         const params ={
             authors : authors,
             book : book
